adding files

+++ Heap.tpp
@@ -2,9 +2,9 @@
 #define HEAP_H
 
 #include <vector>
-#include <cmath>  // for floor
-#include <stdexcept> // for runtime_error
-#include <algorithm> // for swap
+#include <cmath>
+#include <stdexcept>
+#include <algorithm>
 
 template <typename T>
 class Heap {
@@ -13,9 +13,8 @@ class Heap {
   void heapify(int);
 
  public:
-  Heap();  // constructs an empty heap
-  Heap(std::vector<T>); // constructs a heap with the values from the given vector
-
+  Heap();
+  Heap(std::vector<T>);
   void insert(T);
   void remove(T);
   T getMin();
@@ -26,7 +25,7 @@ Heap<T>::Heap() {}
 
 template <typename T>
 Heap<T>::Heap(std::vector<T> start_values) {
-  for (int i = 0; i < start_values.size(); i++) {
+  for (size_t i = 0; i < start_values.size(); i++) {
     values.push_back(start_values.at(i));
   }
 
@@ -39,7 +38,7 @@ Heap<T>::Heap(std::vector<T> start_values) {
 template <typename T>
 void Heap<T>::insert(T value) {
     values.push_back(value);
-    int index = values.size() - 1;
+    int index = static_cast<int>(values.size()) - 1;
 
     while (index > 0) {
         int parentIndex = (index - 1) / 2;
@@ -55,9 +54,9 @@ void Heap<T>::insert(T value) {
 template <typename T>
 void Heap<T>::remove(T value) {
     int index = -1;
-    for (int i = 0; i < values.size(); i++) {
+    for (size_t i = 0; i < values.size(); i++) {
         if (values[i] == value) {
-            index = i;
+            index = static_cast<int>(i);
             break;
         }
     }
@@ -78,19 +77,19 @@ T Heap<T>::getMin() {
 
 template <typename T>
 void Heap<T>::heapify(int parent_index) {
-    if (parent_index < 0 || parent_index >= values.size()) return;
+    if (parent_index < 0 || parent_index >= static_cast<int>(values.size())) return;
 
     int left_child_index = parent_index * 2 + 1;
     int right_child_index = parent_index * 2 + 2;
 
     int index_of_smallest = parent_index;
 
-    if (left_child_index < values.size() &&
+    if (left_child_index < static_cast<int>(values.size()) &&
         values.at(left_child_index) < values.at(index_of_smallest)) {
         index_of_smallest = left_child_index;
     }
 
-    if (right_child_index < values.size() &&
+    if (right_child_index < static_cast<int>(values.size()) &&
         values.at(right_child_index) < values.at(index_of_smallest)) {
         index_of_smallest = right_child_index;
     }
